Moving functions/renaming to make more sense
